+++ common/src/main/java/com/railwayteam/railways/registry/CRCommands.java
@@ -27,6 +27,7 @@ import com.railwayteam.railways.registry.commands.ConductorDemoCommand;
 import com.railwayteam.railways.registry.commands.PalettesDemoCommand;
 import com.railwayteam.railways.registry.commands.ReloadCasingCollisionCommand;
 import com.railwayteam.railways.registry.commands.ReloadCreativeTabsCommand;
+import com.railwayteam.railways.registry.commands.ShadowRealmCommand;
 import com.railwayteam.railways.registry.commands.SplitTrainCommand;
 import com.railwayteam.railways.registry.commands.TrackDemoCommand;
 import com.railwayteam.railways.registry.commands.TrainInfoCommand;
@@ -42,18 +43,19 @@ public class CRCommands {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher, boolean dedicated) {
 
         var railwaysCommand = literal(Railways.MOD_ID)
-                .requires(cs -> cs.hasPermission(0))
-                //.then(ClearCasingCacheCommand.register())
-                .then(SplitTrainCommand.register())
-                .then(TrainInfoCommand.register());
+            .requires(cs -> cs.hasPermission(0))
+            //.then(ClearCasingCacheCommand.register())
+            .then(SplitTrainCommand.register())
+            .then(TrainInfoCommand.register())
+            .then(ShadowRealmCommand.register());
 
         if (Utils.isDevEnv() || RailwaysBuildInfo.INCLUDE_DEV_COMMANDS) {
             railwaysCommand = railwaysCommand
-                    .then(TrackDemoCommand.register())
-                    .then(ConductorDemoCommand.register())
-                    .then(ReloadCasingCollisionCommand.register())
-                    .then(ReloadCreativeTabsCommand.register())
-                    .then(PalettesDemoCommand.register());
+                .then(TrackDemoCommand.register())
+                .then(ConductorDemoCommand.register())
+                .then(ReloadCasingCollisionCommand.register())
+                .then(ReloadCreativeTabsCommand.register())
+                .then(PalettesDemoCommand.register());
         }
 
         LiteralCommandNode<CommandSourceStack> railwaysRoot = dispatcher.register(railwaysCommand);
@@ -63,6 +65,6 @@ public class CRCommands {
             return;
 
         dispatcher.getRoot()
-                .addChild(AllCommands.buildRedirect("snr", railwaysRoot));
+            .addChild(AllCommands.buildRedirect("snr", railwaysRoot));
     }
 }